Initial OSBuild manifest
For now there is only a single manifest to build an image containing most
of default Fedora Workstation package set plus the Phosh shell for mobile.

But it should be easy to manifest for other types of Fedora images and it
could allow to do some experimentation. The existing manifest also adds a
COPR repository that contain custom packages to support the PinePhone Pro.

The images built though could be used in any other device with support for
EFI, and also the manifest allows to build QCOW2 images to import on a VM.

Signed-off-by: Javier Martinez Canillas <javierm@redhat.com>